+++ projects/evo-ui-kit/src/lib/components/evo-table/evo-table-column/evo-table-column.component.ts
@@ -9,6 +9,7 @@ export class EvoTableColumnComponent {
 
     @Input() prop: string = undefined;
     @Input() label = '';
+    @Input() className = '';
     @ContentChild('header', {read: TemplateRef}) header: any;
     @ContentChild('content', {read: TemplateRef}) content: any;